# openapi.yaml
openapi: 3.0.3
info:
  title: NutriCare Connect API
  version: 1.0.0
  description: Local portfolio application API. Seed records are fictional, health thresholds are non-diagnostic, and payment gateways remain simulated. SMTP email can be enabled locally.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /diet-plans:
    post:
      tags: [Diet]
      summary: Create a draft diet plan
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DietPlanRequest' } } } }
      responses: { '201': { description: Draft plan } }
  /diet-plans/{id}/publish:
    post:
      tags: [Diet]
      summary: Publish a plan
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses: { '200': { description: Published plan } }
  /diet-plans/patient/{id}:
    get:
      tags: [Diet]
      summary: List patient plans
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Plans } }
  /progress-logs:
    post:
      tags: [Diet]
      summary: Record meals, water, weight and BMI
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProgressRequest' } } } }
      responses: { '201': { description: Progress log } }
  /progress-logs/patient/{id}:
    get:
      tags: [Diet]
      summary: Return patient trend data
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Progress logs } }
  /messages:
    post:
      tags: [Communication]
      summary: Send a non-urgent secure message
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MessageRequest' } } } }
      responses: { '201': { description: Message } }
  /messages/patient/{id}:
    get:
      tags: [Communication]
      summary: Get patient-care conversation
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Messages } }
  /notifications:
    post:
      tags: [Communication]
      summary: Create an in-app or simulated external notification
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationRequest' } } } }
      responses: { '201': { description: Notification delivery record } }
  /notifications/recipient/{id}:
    get:
      tags: [Communication]
      summary: Get recipient notifications
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Notifications } }
...
  /feedback:
    post:
      tags: [Analytics]
      summary: Save rating and automatically escalate low scores
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FeedbackRequest' } } } }
      responses: { '201': { description: Feedback and optional complaint } }
...
  /reports/summary:
    get:
      tags: [Analytics]
      summary: Return date-filtered management summary
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
      responses: { '200': { description: Aggregate counts and average rating } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    AccountId: { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/AccountId' } }
...
  schemas:
    AccountId: { type: string, pattern: '^[A-Z]{1,2}[0-9]{3}$', example: P001 }
...
    DietPlanRequest:
      type: object
      required: [patientId,title,calorieTarget,mealSchedule]
      description: The plan author ID is derived from the authenticated doctor or dietitian account.
      properties: { patientId: { $ref: '#/components/schemas/AccountId' }, title: { type: string }, calorieTarget: { type: integer }, exclusions: { type: string }, mealSchedule: { type: string } }
    ProgressRequest:
      type: object
      required: [patientId,date]
      properties: { patientId: { $ref: '#/components/schemas/AccountId' }, dietPlanId: { type: string, format: uuid }, date: { type: string, format: date }, weightKg: { type: number }, bmi: { type: number }, waterGlasses: { type: integer }, mealsCompleted: { type: integer } }
    MessageRequest:
      type: object
      required: [senderId,recipientId,patientId,body]
      properties: { senderId: { $ref: '#/components/schemas/AccountId' }, recipientId: { $ref: '#/components/schemas/AccountId' }, patientId: { $ref: '#/components/schemas/AccountId' }, body: { type: string, maxLength: 2000 } }
    NotificationRequest:
      type: object
      required: [recipientId,type,channel,message]
      properties: { recipientId: { $ref: '#/components/schemas/AccountId' }, type: { type: string }, channel: { type: string, enum: [IN_APP,EMAIL,SMS] }, message: { type: string }, simulateFailure: { type: boolean } }
    FeedbackRequest:
      type: object
      required: [patientId,practitionerId,appointmentId,rating]
      properties: { patientId: { $ref: '#/components/schemas/AccountId' }, practitionerId: { $ref: '#/components/schemas/AccountId' }, appointmentId: { type: string, format: uuid }, rating: { type: integer, minimum: 1, maximum: 5 }, comments: { type: string, maxLength: 1500 } }